Bugfix: Adding Multiple Animals to the same spot is now working
I also changed the Test to reflect the changes I made to the Array
holding the Animals. We are no using a 2D Array with an ArrayList as a
container for the Animals. I guess is NOT the best solution, but after
talking to the Prof. She said..."it's for you to see Arrays are not
always the best answer"